requirements.txt
importlib-metadata==6.7.0
Jinja2==3.1.2
MarkupSafe==2.1.3
prettytable==3.7.0
psycopg2==2.9.7
python-dotenv==0.21.1
typing_extensions==4.7.1
wcwidth==0.2.6
zipp==3.15.0
